uihugh

+++ CEA.Domain/Constants/ViaticoImporte.cs
@@ -3,14 +3,14 @@ namespace CEA.Domain.Constants
 {
     public static class ViaticoImporte
     {
-        public const decimal ImporteViaticoDirectorDentroEstado = 350;
-        public const decimal ImporteViaticoDirectorFueraEstado = 600;
+        public const double ImporteViaticoDirectorDentroEstado = 350;
+        public const double ImporteViaticoDirectorFueraEstado = 600;
 
-        public const decimal ImporteViaticoJefeDentroEstado = 300;
-        public const decimal ImporteViaticoJefeFueraEstado = 500;
+        public const double ImporteViaticoJefeDentroEstado = 300;
+        public const double ImporteViaticoJefeFueraEstado = 500;
 
-        public const decimal ImporteViaticoEmpleadoDentroEstado = 270;
-        public const decimal ImporteViaticoEmpleadoFueraEstado = 400;
+        public const double ImporteViaticoEmpleadoDentroEstado = 270;
+        public const double ImporteViaticoEmpleadoFueraEstado = 400;
 
 
     }
